fix: move base api client and test case logic
